🐛 fix : 전체적인 css 버그 수정
