+++ src/app/(with-header)/myprofile/_components/MyReviewListContainer.tsx
@@ -1,50 +1,17 @@
 'use client';
 
 import Image from 'next/image';
-import { useEffect, useState } from 'react';
-import elapsedTime from '@/utils/formatDate';
-import like from '@/assets/icons/star_hover.svg';
+import { useCallback, useEffect, useState } from 'react';
 import { fetchWithAuth } from '@/lib/auth';
 import emptyData from '@/assets/icons/empty_review.svg';
 import { MyReview, MyReviewResponse } from '@/types/review-data';
-import KebabDropDown from './KebabDropDown';
-
-interface MyReviewProps {
-  id: number;
-  rating: number;
-  createdAt: string;
-  wineName: string;
-  content: string;
-}
-
-export function MyReviewItem({ id, rating, createdAt, wineName, content }: MyReviewProps) {
-  const reviewElapsedTime = elapsedTime(createdAt);
-
-  return (
-    <div className='flex w-[800px] rounded-[16px] border border-gray-300 px-10 py-7 tablet:w-full mobile:w-full'>
-      <div className='flex flex-grow flex-col gap-[20px]'>
-        <div className='relative flex gap-[15px]'>
-          <div className='flex h-[42px] w-[80px] items-center justify-center gap-[4px] rounded-[12px] bg-purple-10 tablet:h-[38px] mobile:h-[32px] mobile:w-[60px]'>
-            <Image className='w-[20px] mobile:w-[16px]' src={like} alt='별점 아이콘' />
-            <span className='text-[18px] font-bold text-purple-100 mobile:text-[14px]'>{rating.toFixed(1)}</span>
-          </div>
-          <KebabDropDown id={id} />
-          <span className='flex items-center justify-center text-[16px] text-gray-500 mobile:text-[14px]'>{reviewElapsedTime}</span>
-        </div>
-        <div className='flex flex-col gap-[10px]'>
-          <h3 className='text-[16px] font-medium text-gray-500 mobile:text-[14px]'>{wineName}</h3>
-          <p className='text-[16px] font-normal text-gray-800 mobile:text-[14px]'>{content}</p>
-        </div>
-      </div>
-    </div>
-  );
-}
+import { MyReviewItem } from './MyReviewItem';
 
 export default function MyReviewListContainer({ setDataCount }: { setDataCount: (value: number) => void }) {
   const [myReviewData, setMyReviewData] = useState<MyReview[]>([]);
   const [isLoading, setIsloading] = useState(true);
 
-  const getMyReview = async () => {
+  const getMyReview = useCallback(async () => {
     try {
       setIsloading(true);
       const response = await fetchWithAuth(`${process.env.NEXT_PUBLIC_BASE_URL}/users/me/reviews?limit=30`);
@@ -55,17 +22,17 @@ export default function MyReviewListContainer({ setDataCount }: { setDataCount: 
 
       const data: MyReviewResponse = await response.json();
       setMyReviewData(data.list);
+      setDataCount(data.totalCount);
     } catch (error) {
       console.error('데이터를 불러오는데 오류가 발생했습니다:', error);
     } finally {
       setIsloading(false);
     }
-  };
+  }, [setDataCount]);
 
   useEffect(() => {
     getMyReview();
-    setDataCount(myReviewData.length);
-  }, [setDataCount, myReviewData.length]);
+  }, [getMyReview]);
 
   if (isLoading) return <div></div>;
